Added Nuked OPL3 emulator (which is more accurate than DosBox's) Also:  - Fixed warnings in the CLang code model plugin for Qt Creator  - Fixed bend coefficient which caused incorrect hi-hats in DMX banks

+++ src/adlmidi.h
@@ -28,7 +28,8 @@
 extern "C" {
 #endif
 
-struct ADL_MIDIPlayer {
+struct ADL_MIDIPlayer
+{
     unsigned int AdlBank;
     unsigned int NumFourOps;
     unsigned int NumCards;
@@ -49,9 +50,9 @@ struct ADL_MIDIPlayer {
     double maxdelay;
 
     /*For internal usage*/
-    int stored_samples; //num of collected samples
-    int backup_samples[1024]; //Backup sample storage.
-    int backup_samples_size; //Backup sample storage.
+    int     stored_samples; //num of collected samples
+    short   backup_samples[1024]; //Backup sample storage.
+    int     backup_samples_size; //Backup sample storage.
     /*For internal usage*/
 
     void *adl_midiPlayer;
@@ -59,62 +60,61 @@ struct ADL_MIDIPlayer {
 };
 
 /* Sets number of emulated sound cards (from 1 to 100). Emulation of multiple sound cards exchanges polyphony limits*/
-extern int adl_setNumCards(struct ADL_MIDIPlayer*device, int numCards);
+extern int adl_setNumCards(struct ADL_MIDIPlayer *device, int numCards);
 
 /* Sets a number of the patches bank from 0 to N banks */
-extern int adl_setBank(struct ADL_MIDIPlayer* device, int bank);
+extern int adl_setBank(struct ADL_MIDIPlayer *device, int bank);
 
 /* Returns total number of available banks */
 extern int adl_getBanksCount();
 
 /* Returns pointer to array of names of every bank */
-extern const char * const* adl_getBankNames();
+extern const char *const *adl_getBankNames();
 
 /*Sets number of 4-chan operators*/
-extern int adl_setNumFourOpsChn(struct ADL_MIDIPlayer*device, int ops4);
+extern int adl_setNumFourOpsChn(struct ADL_MIDIPlayer *device, int ops4);
 
 /*Enable or disable AdLib percussion mode*/
-extern void adl_setPercMode(struct ADL_MIDIPlayer* device, int percmod);
+extern void adl_setPercMode(struct ADL_MIDIPlayer *device, int percmod);
 
 /*Enable or disable deep vibrato*/
-extern void adl_setHVibrato(struct ADL_MIDIPlayer* device, int hvibro);
+extern void adl_setHVibrato(struct ADL_MIDIPlayer *device, int hvibro);
 
 /*Enable or disable deep tremolo*/
-extern void adl_setHTremolo(struct ADL_MIDIPlayer* device, int htremo);
+extern void adl_setHTremolo(struct ADL_MIDIPlayer *device, int htremo);
 
 /*Enable or disable Enables scaling of modulator volumes*/
-extern void adl_setScaleModulators(struct ADL_MIDIPlayer* device, int smod);
+extern void adl_setScaleModulators(struct ADL_MIDIPlayer *device, int smod);
 
 /*Enable or disable built-in loop (built-in loop supports 'loopStart' and 'loopEnd' tags to loop specific part)*/
-extern void adl_setLoopEnabled(struct ADL_MIDIPlayer* device, int loopEn);
+extern void adl_setLoopEnabled(struct ADL_MIDIPlayer *device, int loopEn);
 
 /*Enable or disable Logariphmic volume changer */
-extern void adl_setLogarithmicVolumes(struct ADL_MIDIPlayer* device, int logvol);
+extern void adl_setLogarithmicVolumes(struct ADL_MIDIPlayer *device, int logvol);
 
 /*Returns string which contains last error message*/
-extern const char* adl_errorString();
+extern const char *adl_errorString();
 
 /*Initialize ADLMIDI Player device*/
-extern struct ADL_MIDIPlayer* adl_init(long sample_rate);
+extern struct ADL_MIDIPlayer *adl_init(long sample_rate);
 
 /*Load MIDI file from File System*/
-extern int adl_openFile(struct ADL_MIDIPlayer* device, char *filePath);
+extern int adl_openFile(struct ADL_MIDIPlayer *device, char *filePath);
 
 /*Load MIDI file from memory data*/
-extern int adl_openData(struct ADL_MIDIPlayer* device, void* mem, long size);
+extern int adl_openData(struct ADL_MIDIPlayer *device, void *mem, long size);
 
 /*Resets MIDI player*/
-extern void adl_reset(struct ADL_MIDIPlayer*device);
+extern void adl_reset(struct ADL_MIDIPlayer *device);
 
 /*Close and delete ADLMIDI device*/
 extern void adl_close(struct ADL_MIDIPlayer *device);
 
 /*Take a sample buffer*/
-extern int  adl_play(struct ADL_MIDIPlayer*device, int sampleCount, short out[]);
+extern int  adl_play(struct ADL_MIDIPlayer *device, int sampleCount, short out[]);
 
 #ifdef __cplusplus
-     }
+}
 #endif
 
 #endif // ADLMIDI_H
-
